004 create OrderRequestDto and fixed db

+++ server/mobile-shop/src/main/java/sit/int204/mobileshop/dtos/OrderItemDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @JsonPropertyOrder({ "no", "saleItemId", "price", "quantity", "description" })
 public class OrderItemDto {
     private Integer no;           
-    private Long saleItemId;
+    private Integer saleItemId;
     private Integer price;
     private Integer quantity;
     private String description;
